requirements.txt
allure-pytest==2.12.0
allure-python-commons==2.12.0
pytest==7.2.1
pytest-order==1.0.1
pytest-ordering==0.6
requests==2.28.2